feat(lib): universal header component
Replaces headers in all apps with new universal header component.
