--- apps/admin-gui/src/app/shared/perun-nav/perun-nav.component.html
@@ -1,38 +1,0 @@
-<div [ngStyle]="{'color': navTextColor}" id="nav-menu-icons">
-  <button (click)="sideNav.toggle()" class="mr-2 mt-auto mb-auto" mat-icon-button>
-    <mat-icon [ngStyle]="{'color': iconColor}"> menu </mat-icon>
-  </button>
-  <a [routerLink]="['/home']" class="mt-auto mb-auto" queryParamsHandling="merge">
-    <div [innerHTML]="logo" class="logo-container"></div>
-  </a>
-</div>
-
-<div *ngIf="isDevel" id="nav-devel-info">
-  <span class="text-warning"> ~ 🗲&nbsp; </span>
-  <span class="text-danger"> DEVEL </span>
-  <span class="text-warning"> &nbsp;🗲 ~ </span>
-</div>
-
-<div [ngStyle]="{'color': navTextColor}" id="nav-menu-user-info">
-  <button
-    [matTooltip]="'NAV.NOTIFICATIONS_TOOLTIP'|translate"
-    (click)="showNotificationHistory()"
-    mat-icon-button>
-    <mat-icon
-      [ngStyle]="{'color': iconColor}"
-      matBadge="{{getNewNotificationsCount() > 5 ? '5+' : getNewNotificationsCount()}}"
-      [matBadgeHidden]="getNewNotificationsCount() === 0"
-      matBadgeColor="warn">
-      {{getNewNotificationsCount() ? 'notifications_active' : 'notifications'}}
-    </mat-icon>
-  </button>
-
-  <perun-web-apps-header-menu
-    [iconColor]="iconColor"
-    [label]="profileLabel"
-    [textColor]="navTextColor"
-    [logoutEnabled]="logoutEnabled"
-    [url]="profileUrl"
-    [user]="principal.user">
-  </perun-web-apps-header-menu>
-</div>
