--- apps/admin-gui/src/app/shared/perun-nav/perun-nav.component.scss
@@ -1,42 +1,0 @@
-#nav-menu-user-info {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-#nav-devel-info {
-  font-weight: 500;
-  font-size: 1.4rem;
-  display: flex;
-  align-items: center;
-  font-family: Roboto, 'Helvetica Neue', sans-serif;
-}
-
-#nav-menu-icons {
-  margin-left: 8px;
-  display: flex;
-  flex-direction: row;
-  width: 200px;
-}
-
-.logo-container {
-  padding: 0 8px 0 8px;
-}
-
-.logo-container ::ng-deep img,
-.logo-container ::ng-deep svg {
-  max-height: 50px;
-}
-
-#nav-link-items {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  & a {
-    padding: 0 1rem;
-    line-height: 64px;
-    text-align: center;
-    height: 64px;
-    display: block;
-  }
-}
