+++ apps/admin-gui/src/assets/config/defaultConfig.json
@@ -8,7 +8,6 @@
   "supported_languages": [
     "en"
   ],
-  "is_devel": false,
   "display_warning": false,
   "warning_message": "BETA version of new Perun GUI. It doesn’t contain all functionality. Please report any issues to support mail below.",
   "auto_auth_redirect": true,
