+++ apps/consolidator/src/app/app.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { InitAuthService, StoreService } from '@perun-web-apps/perun/services';
+import { AppType } from '@perun-web-apps/perun/models';
 
 @Component({
   selector: 'perun-web-apps-root',
@@ -9,6 +10,7 @@ import { InitAuthService, StoreService } from '@perun-web-apps/perun/services';
 export class AppComponent {
   contentBackgroundColor = this.store.getProperty('theme').content_bg_color;
   isLoginScreenShow = this.initAuth.isLoginScreenShown();
+  otherApp = AppType.Profile;
 
   constructor(private store: StoreService, private initAuth: InitAuthService) {}
 }
