--- apps/consolidator/src/app/components/header/header.component.css
@@ -1,41 +1,0 @@
-#nav-devel-info {
-  margin-left: auto;
-  font-weight: 500;
-  font-size: 1.4rem;
-  display: flex;
-  align-items: center;
-  font-family: Roboto, 'Helvetica Neue', sans-serif;
-}
-
-#login-nav {
-  position: fixed;
-  width: 100%;
-  z-index: 999;
-  display: flex;
-  height: 64px;
-  max-height: 64px;
-  top: 0;
-}
-
-#nav-menu-user-info {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding-right: 16px;
-  margin-left: auto;
-  margin-right: 0;
-}
-
-.logo-container {
-  padding: 0 8px 0 8px;
-}
-
-.logo-container ::ng-deep img,
-.logo-container ::ng-deep svg {
-  max-height: 50px;
-}
-
-.page-subtitle {
-  margin-bottom: 1rem;
-  font-size: 1.5rem;
-}
