--- apps/consolidator/src/app/components/header/header.component.html
@@ -1,24 +1,0 @@
-<div [ngStyle]="{ background: bgColor, color: textColor }" id="login-nav" class="mat-elevation-z3">
-  <a [routerLink]="['/consolidate']" class="mt-auto mb-auto" queryParamsHandling="merge">
-    <div [innerHTML]="logo" class="logo-container"></div>
-  </a>
-  <p [ngStyle]="{ color: textColor }" class="ml-3 mt-auto mb-auto page-subtitle">
-    {{ 'HEADER.TITLE' | translate }}
-  </p>
-
-  <div *ngIf="isDevel" id="nav-devel-info">
-    <span class="text-warning"> ~ 🗲&nbsp; </span>
-    <span class="text-danger"> DEVEL </span>
-    <span class="text-warning"> &nbsp;🗲 ~ </span>
-  </div>
-
-  <perun-web-apps-header-menu
-    id="nav-menu-user-info"
-    [iconColor]="iconColor"
-    [label]="profileLabel"
-    [textColor]="textColor"
-    [logoutEnabled]="logoutEnabled"
-    [url]="profileUrl"
-    [user]="principal.user">
-  </perun-web-apps-header-menu>
-</div>
