+++ apps/password-reset/src/app/app.module.ts
@@ -20,7 +20,6 @@ import { PasswordResetConfigService } from './services/password-reset-config.ser
 import { PERUN_API_SERVICE } from '@perun-web-apps/perun/tokens';
 import { UiMaterialModule } from '@perun-web-apps/ui/material';
 import { OAuthModule, OAuthStorage } from 'angular-oauth2-oidc';
-import { HeaderComponent } from './components/header/header.component';
 import { PasswordResetPageComponent } from './pages/password-reset-page/password-reset-page.component';
 import { PasswordResetFormComponent } from './components/password-reset-form/password-reset-form.component';
 import { UiAlertsModule } from '@perun-web-apps/ui/alerts';
@@ -52,7 +51,6 @@ const loadConfigs = (appConfig: PasswordResetConfigService) => (): Promise<void>
 @NgModule({
   declarations: [
     AppComponent,
-    HeaderComponent,
     PasswordResetPageComponent,
     PasswordResetFormComponent,
     InvalidRequestAlertComponent,
