--- apps/password-reset/src/app/components/header/header.component.html
@@ -1,25 +1,0 @@
-<mat-toolbar
-  [ngStyle]="{'background': bgColor}"
-  class="sticky-top header-container mat-elevation-z3">
-  <div class="display-flex-row">
-    <a [innerHTML]="logo" class="logo-container mt-auto mb-auto"></a>
-    <p [ngStyle]="{'color': textColor}" class="ml-3 mt-auto mb-auto">{{label}}</p>
-  </div>
-
-  <div *ngIf="isDevel" id="nav-devel-info">
-    <span class="text-warning"> ~ 🗲&nbsp; </span>
-    <span class="text-danger"> DEVEL </span>
-    <span class="text-warning"> &nbsp;🗲 ~ </span>
-  </div>
-
-  <div id="nav-menu-language" [ngStyle]="{'color': textColor}">
-    <button mat-button [matMenuTriggerFor]="menu">
-      <mat-icon [ngStyle]="{'color': iconColor}"> language </mat-icon>
-    </button>
-    <mat-menu #menu="matMenu">
-      <button mat-menu-item (click)="changeLanguage()">
-        {{'HEADER.LANGUAGE' | translate}}
-      </button>
-    </mat-menu>
-  </div>
-</mat-toolbar>
