+++ apps/publications/src/app/app.component.html
@@ -1,6 +1,6 @@
 <perun-web-apps-notificator></perun-web-apps-notificator>
 <div *ngIf="!isLoginScreenShow && !isServiceAccess && !isServiceLogin()">
-  <perun-web-apps-header [sideNav]="sidenav" id="nav-menu"></perun-web-apps-header>
+  <perun-web-apps-header (sidenavToggle)="sidenav.toggle()"></perun-web-apps-header>
   <mat-sidenav-container autosize="true">
     <mat-sidenav
       #sidenav
