--- apps/publications/src/app/components/header/header.component.html
@@ -1,42 +1,0 @@
-<mat-toolbar [ngStyle]="{'background': bgColor}" class="header-container mat-elevation-z3">
-  <div class="display-flex">
-    <button
-      (click)="sideNav.toggle()"
-      [ngStyle]="{'color': iconColor}"
-      mat-icon-button
-      class="mt-auto mb-auto">
-      <mat-icon>menu</mat-icon>
-    </button>
-
-    <a [innerHTML]="logo" routerLink="/all-publications" class="logo-container mt-auto mb-auto"></a>
-    <mat-icon
-      svgIcon="perun-publications-white"
-      class="mt-auto mb-auto router-icon ml-2 mr-4"
-      routerLink="/all-publications"></mat-icon>
-    <p [ngStyle]="{'color': textColor}" class="mt-auto mb-auto page-subtitle">
-      {{'HEADER.TITLE' | translate}}
-    </p>
-  </div>
-
-  <div *ngIf="isDevel" id="nav-devel-info">
-    <span class="text-warning"> ~ 🗲&nbsp; </span>
-    <span class="text-danger"> DEVEL </span>
-    <span class="text-warning"> &nbsp;🗲 ~ </span>
-  </div>
-
-  <div id="nav-menu-user-info" [ngStyle]="{'color': textColor}">
-    <button mat-button [matMenuTriggerFor]="menu">
-      <mat-icon [ngStyle]="{'color': iconColor}"> person </mat-icon>
-      <span class="d-none d-lg-inline">
-        {{principal.user.titleBefore}}
-        {{principal.user.firstName}}
-        {{principal.user.middleName}}
-        {{principal.user.lastName}}
-        {{principal.user.titleAfter}}
-      </span>
-    </button>
-    <mat-menu #menu="matMenu">
-      <button mat-menu-item (click)="onLogOut()">{{'HEADER.LOGOUT' | translate}}</button>
-    </mat-menu>
-  </div>
-</mat-toolbar>
