+++ apps/publications/src/assets/config/defaultConfig.json
@@ -19,6 +19,7 @@
   "supported_languages": [
     "en"
   ],
+  "header_label_en": "Publications",
   "auto_auth_redirect": true,
   "footer": {
     "columns": [
@@ -97,5 +98,6 @@
     "sidemenu-link-active": "#465258",
     "sidemenu_active_text_color": "#ffffff",
     "user_color": "#00796b"
-  }
+  },
+  "log_out_enabled": true
 }
