+++ apps/publications/src/assets/i18n/en.json
@@ -531,5 +531,8 @@
         "NO_USERS": "No users found."
       }
     }
+  },
+  "NAV": {
+    "LOGOUT": "Log out"
   }
 }
