+++ apps/user-profile/src/app/app.component.html
@@ -1,7 +1,9 @@
 <ng-template [ngIf]="!isLoginScreenShown && !isServiceAccess && !isServiceLogin()">
   <perun-web-apps-notificator></perun-web-apps-notificator>
   <div class="app-min-width">
-    <perun-web-apps-header (sidenavToggle)="sidenav.toggle()"></perun-web-apps-header>
+    <perun-web-apps-header
+      (sidenavToggle)="sidenav.toggle()"
+      [otherApp]="otherApp"></perun-web-apps-header>
     <mat-sidenav-container autosize="true">
       <mat-sidenav
         #sidenav
