--- apps/user-profile/src/app/components/header/header.component.html
@@ -1,38 +1,0 @@
-<mat-toolbar
-  [ngStyle]="{'background': bgColor}"
-  class="sticky-top header-container mat-elevation-z3">
-  <div id="nav-menu-icons">
-    <button
-      (click)="onToggleSidenav()"
-      *ngIf="!hideToggle"
-      [ngStyle]="{'color': iconColor}"
-      aria-label="Menu"
-      mat-icon-button>
-      <mat-icon>menu</mat-icon>
-    </button>
-
-    <a
-      [innerHTML]="logo"
-      [routerLink]="disableLogo ? [] : ['/']"
-      class="logo-container mt-auto mb-auto"
-      queryParamsHandling="merge"></a>
-
-    <p [ngStyle]="{'color': textColor}" class="ml-3 mt-auto mb-auto hide-label">{{label}}</p>
-  </div>
-
-  <div *ngIf="isDevel" id="nav-devel-info">
-    <span class="text-warning"> ~ 🗲&nbsp; </span>
-    <span class="text-danger"> DEVEL </span>
-    <span class="text-warning"> 🗲 ~ </span>
-  </div>
-
-  <perun-web-apps-header-menu
-    id="nav-menu-user-info"
-    [url]="adminGuiUrl"
-    [label]="adminLabel"
-    [user]="principal.user"
-    [logoutEnabled]="logoutEnabled"
-    [iconColor]="iconColor"
-    [textColor]="textColor">
-  </perun-web-apps-header-menu>
-</mat-toolbar>
