--- apps/user-profile/src/app/components/header/header.component.scss
@@ -1,43 +1,0 @@
-.header-container {
-  height: 64px !important;
-  padding: 0 0 0 8px;
-  z-index: 999;
-}
-
-#nav-menu-user-info {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding-right: 16px;
-  margin-left: auto;
-  margin-right: 0;
-}
-
-@media (max-width: 570px) {
-  .hide-label {
-    display: none;
-  }
-}
-
-#nav-devel-info {
-  margin-left: auto;
-  font-weight: 500;
-  font-size: 1.4rem;
-  display: flex;
-  align-items: center;
-  font-family: Roboto, 'Helvetica Neue', sans-serif;
-}
-
-#nav-menu-icons {
-  display: flex;
-  flex-direction: row;
-}
-
-.logo-container {
-  padding: 0 8px 0 8px;
-}
-
-.logo-container ::ng-deep img,
-.logo-container ::ng-deep svg {
-  max-height: 50px;
-}
