+++ libs/perun/components/src/lib/perun-header/perun-header.component.html
@@ -1,0 +1,61 @@
+<mat-toolbar
+  [ngStyle]="{'background': bgColor}"
+  class="sticky-top header-container mat-elevation-z3">
+  <div id="nav-menu-icons">
+    <button
+      (click)="onToggleSidenav()"
+      *ngIf="showToggle"
+      [ngStyle]="{'color': iconColor}"
+      aria-label="Menu"
+      mat-icon-button>
+      <mat-icon>menu</mat-icon>
+    </button>
+
+    <a
+      [innerHTML]="logo"
+      [routerLink]="disableLogo ? [] : ['/']"
+      class="logo-container mt-auto mb-auto"
+      queryParamsHandling="merge"></a>
+
+    <p [ngStyle]="{'color': textColor}" class="mt-auto mb-auto hide-label page-subtitle">
+      {{label}}
+    </p>
+  </div>
+
+  <div [ngStyle]="{'color': navTextColor}" id="nav-menu-user-info">
+    <div id="nav-menu-language" [ngStyle]="{'color': textColor}" *ngIf="showLanguageMenu">
+      <button mat-button [matMenuTriggerFor]="menu">
+        <mat-icon [ngStyle]="{'color': iconColor}"> language </mat-icon>
+      </button>
+      <mat-menu #menu="matMenu">
+        <button mat-menu-item (click)="changeLanguage()">
+          {{'HEADER.LANGUAGE' | translate}}
+        </button>
+      </mat-menu>
+    </div>
+
+    <button
+      *ngIf="showNotifications"
+      [matTooltip]="'NAV.NOTIFICATIONS_TOOLTIP'|translate"
+      (click)="showNotificationHistory()"
+      mat-icon-button>
+      <mat-icon
+        [ngStyle]="{'color': iconColor}"
+        matBadge="{{getNewNotificationsCount() > 5 ? '5+' : getNewNotificationsCount()}}"
+        [matBadgeHidden]="getNewNotificationsCount() === 0"
+        matBadgeColor="warn">
+        {{getNewNotificationsCount() ? 'notifications_active' : 'notifications'}}
+      </mat-icon>
+    </button>
+
+    <perun-web-apps-header-menu
+      *ngIf="showHeaderMenu"
+      [url]="otherAppUrl"
+      [label]="otherAppLabel"
+      [user]="principal.user"
+      [logoutEnabled]="logoutEnabled"
+      [iconColor]="iconColor"
+      [textColor]="textColor">
+    </perun-web-apps-header-menu>
+  </div>
+</mat-toolbar>
