+++ libs/perun/dialogs/src/lib/show-notification-history-dialog/show-notification-history-dialog.component.html
@@ -1,4 +1,4 @@
-<h1 mat-dialog-title>{{'DIALOGS.NOTIFICATION_LIST.TITLE' | translate}}</h1>
+<h1 mat-dialog-title>{{'DIALOGS.NOTIFICATION_LIST.TITLE' | customTranslate | translate}}</h1>
 <div *ngIf="notifications.length !== 0" class="hide-horizontal-scrollbar">
   <cdk-virtual-scroll-viewport class="viewport" itemSize="50">
     <div *cdkVirtualFor="let data of notifications" class="mb-1">
@@ -9,14 +9,14 @@
 <perun-web-apps-alert
   alert_type="warn"
   *ngIf="notifications.length === 0"
-  >{{'DIALOGS.NOTIFICATION_LIST.NO_NOTIFICATIONS_ALERT'|translate}}</perun-web-apps-alert
+  >{{'DIALOGS.NOTIFICATION_LIST.NO_NOTIFICATIONS_ALERT'| customTranslate | translate}}</perun-web-apps-alert
 >
 
 <div mat-dialog-actions>
   <button (click)="onClear()" class="ml-auto" mat-flat-button>
-    {{'DIALOGS.NOTIFICATION_LIST.CLEAR' | translate}}
+    {{'DIALOGS.NOTIFICATION_LIST.CLEAR' | customTranslate | translate}}
   </button>
   <button (click)="onCancel()" class="ml-2" mat-flat-button>
-    {{'DIALOGS.NOTIFICATION_LIST.CANCEL' | translate}}
+    {{'DIALOGS.NOTIFICATION_LIST.CANCEL' | customTranslate | translate}}
   </button>
 </div>
