+++ libs/perun/models/src/lib/AppType.ts
@@ -1,0 +1,7 @@
+export enum AppType {
+  Admin = 'admin',
+  Profile = 'profile',
+  PwdReset = 'pwdReset',
+  Consolidator = 'consolidator',
+  Linker = 'linker',
+}
