+++ libs/perun/models/src/lib/ConfigProperties.ts
@@ -149,7 +149,6 @@ export interface PerunConfig {
   auto_auth_redirect?: boolean;
   supported_languages?: string[];
   enforce_consents?: boolean;
-  is_devel?: boolean;
   instance_favicon?: boolean;
   log_out_enabled?: boolean;
   header_label_en?: string;
